controller: Use timing API for OVS/SB updates.
This RFC patch enables batch processing of OVS and southbound database
updates using the _until variant of the ovsdb_idl_loop_run() function
proposed in the patch set at [1] and using a timeout of 500ms.  It also
adds new stopwatches to allow operators to observe how much time is
actually spent in the idl loop.

[1] https://mail.openvswitch.org/pipermail/ovs-dev/2025-May/423275.html

Signed-off-by: Martin Morgenstern <martin.morgenstern@cloudandheat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>