Datapath_Binding: Separate type and UUID external-ids.
Prior to this commit, if you wanted to find the corresponding northbound
UUID for a southbound Logical Datapath, you could find it in one of two
places:

For logical switches, it was in external-ids:logical-switch.
For logical routers, it was in external-ids:logical-router.

With this commit, we are changing the external-ids so that the
northbound UUID and the type correspond to separate keys. The
external-ids:type corresponds to an enum value for logical routers and
logical switches. The external-ids:nb_uuid gets the northbound UUID of
the logical switch or logical router. And for human readability, we also
are providing a external-ids:type_str that will be "logical-router" for
logical routers and "logical-switch" for logical switches.

Note that there is a seemingly-unrelated change in the check packet
length test in tests/ovn.at. This test started failing because its use
of `grep` was picking up the new "nb_uuid" external-id accidentally. The
change to the test uses `fetch_column` since it is more precise.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>