controller: Prevent assert with BGP when SB is read only.
There wasn't any check if SB is writable when we tried to commit
learned routes into SB, this would lead to assert and crash. In
order to prevent that track the SB state and add additional handler
that will trigger recompute of the route_exchange node if there are
SB changes pending and the SB is writable again.

Fixes: 866a5014ae45 ("controller: Support learning routes.")
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>