ovs: Update ovs submodule commit.
This updates the submodule to include a change to db-ctl-base.c that
allows for columns of type UUID to be used as record IDs when retrieving
database rows using the ctl_get_row() function.

The submodule update actually goes beyond the above-referenced commit in
order to allow for EVPN development that is forthcoming.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>