northd: Process port_group changes incrementally
Several operations on Port Groups required a full database
recalculation.

1. Creating a new Port Group
2. Deleting a Port Group
3. adding a port from a switch with no other ports on a port group
4. removing a port from a port group that is the last one from a
   specific switch

Those four operations required a full database recalculation to ensure
that that the lflows from the ACLs and the southbound databases where
correctly populated. Instead reprocess only the relevent switches and
ACLs to avoid the cost of a full recalculation. This is done by having
the port_group_handler update the southbound database and pass the names
of the relevent switches to the ls_stateful_acl node.

Reported-at: https://issues.redhat.com/browse/FDP-758
Reported-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>