tests: Add support for running system tests under retis.
Retis is very useful for debugging our system tests or debugging
kernel issues through our system tests.  This change adds a convenient
way to run any kernel system test with the retis capture on the
background.  E.g.:

  make check-kernel OVS_TEST_WITH_RETIS=yes TESTSUITEFLAGS='168 -d'

Retis 1.5 is required, since we're using ifdump profile, and it also
will mount debugfs for us in case of running in a different namespace.
It should be available in $PATH.

In addition to just capturing the retis.data, we're also running the
capture with --print to print all the events as they appear, and
producing the sorted output in the end.  This makes it easier to work
across systems with different versions of retis and saves time for
running the sort manually.  The raw data is still available for
advanced processing, if needed.

Not specifying any particular collector, capturing everything that's
enabled by default.  OVS tracking is turned on by default.

Since OVS tracking is used, it's required to start retis after the
kernel datapath is created, otherwise it will fail to obtain the map
of upcall PIDs.  That's why we need to start it after the bridge is
created.  This also makes it necessary to build the OVS submodule
with --enable-usdt-probes flag, otherwise retis may refuse to start
when it can't find any probes in the running ovs-vswitchd.

Only adding support for kernel test suite for now.  For userspace test
suites it may also be useful at some point, but currently that requires
running without --ovs-track and isn't too important.

Startup of the retis capture adds significant amount of time to each
test, so not running it by default.

This change is ported from the corresponding OVS change:
  https://github.com/openvswitch/ovs/commit/22732c0e6770597031d1eb0a61ec922ea296c8b1
The environment variable name starts with OVS_ for that reason.  But
we already have the OVS_PAUSE_TEST, so just following the pattern.

The DAEMONIZE() macro is changed to use 'on_exit' instead of appending
commands directly to the 'cleanup' file.  That ensures the correct
ordering between stopping the daemon and running the 'retis sort', as
'on_exit' ensures that cleanup commands are executed in the reverse
order.  This makes the macro similar to OVS_DAEMONIZE in the OVS repo.
In the future, a similar change can be made in the NETNS_DAEMONIZE()
as well to ensure the proper ordering, but it's not necessary for now.

Link: https://github.com/retis-org/retis
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>