controller: Fix missing garp while gw are fighting.
When two gateways fight in ha mode to bind the router port,
there were cases in which one of the gateway almost always
immediately reclaimed the router port, as the grace time was
honored by the other gw. This is also the case since [1] for
the highest priority gw.
In such a case, the garp_rarp module did not detect the change in
the router port binding (as the port bound to a different chassis
in idl was already rebound in runtime_data, potentially during a
recompute.
Hence it was not resetting the garp "count", potentially resulting
in no new garps.

Also, with a fight time of 500 msec there is not enough time for gw
to send garp, so increase the grace time to two seconds.

[1] controller: Do not postpone claim for highest priority chassis.

Fixes: 8ab8570e21ff ("pinctrl: Fix missing garp.")
Reported-at: https://issues.redhat.com/browse/FDP-1418
Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>