northd: Do not advertise NAT and LBs when "connected-as-host" is configured.
When "connected-as-host" was configured for option
"dynamic-routing-redistribute" OVN would attempt to advertise also
NATs and LBs connected to that LRP. Prevent that by advertising only
the LRP networks. This also prevents commit fail loop when
"connected-as-host" and "nat" are used at the same time. Also fix
the case when there is "conflict" between LRP IP and SNAT. In
that case honor the LRP IP advertised route.

Fixes: 93f541f342f9 ("northd: Allow announcing individual host routes.")
Reported-at: https://issues.redhat.com/browse/FDP-1564
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>