release-process: Add time between soft freeze and branching.
OVN has grown in popularity over the years, and as a result, we are
receiving many more feature request patches than we had in previous
release cycles. The past two or three release cycles, we have had to
either delay soft freeze or extend soft freeze in order to allow for
requested submissions to be considered.

This documentation change proposes extending the soft freeze period from
two weeks to four weeks, to give more time for reviews prior to testing.

From the release calendar update in 2026, you can see that the way this
is accomplished is by moving the branch and subsequent release two weeks
later than they had previously been. The other option was to move soft
freeze two weeks earlier, but this has the potential to cause problems
for patches that rely on new OVS features. OVS soft freeze currently
ends right as OVN's is beginning. Thus if an OVN feature relies on an
OVS patch to go in first, then it works out well with the current
schedule. If OVN and OVS have overlapping soft freeze periods, then
trying to get the OVN patches reviewed could be a challenge.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>