northd: Create cachable pointer for datapath bindings.
Calling sbrec_datapath_binding_insert() returns a pointer to the
datapath binding that will be inserted once the transaction is
committed. This pointer is only valid until the point the transaction is
committed. At that point, the IDL will free the pointer and it is no
longer valid. Caching this pointer, therefore, is dangerous, since you
could be holding a pointer to freed data. Once the IDL runs again, it
will provide a permanent pointer to the datapath binding, which is safe
to cache.

This patch introduces a wrapper structure, ovn_datapath_binding. This
wrapper structure is safe to cache since it is not deleted by the IDL.
The inner sbrec_datapath_binding pointer can be updated by datapath
syncing nodes when the IDL provides a new datapath binding pointer.

This particular patch does nothing other than to introduce the structure
and cache it where necessary. Upcoming commits will cause datapath
syncing nodes to update their inner pointers as necessary, making
downstream engine node caching safe.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>