controller: Pair remote VTEPs with datapaths.
Pair learned remote VTEPs with their respective datapaths and tunnel
ports. This creates binding that can be used for physical flows. In
order to accommodate for multicast groups we need to also create
specific EVPN multicast groups.

All data are ovn-controller specific due to the nature of EVPN and
also to avoid possible scalability issues within SB database.

Those bindings and multicast groups are used in later commits to
create physical flows responsible for the delivery.

Note that there isn't any way to set vni for given LS that will
be part of future patch ("dynamic-routing-vni").

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>