controller: Create physical flows based on the advertised EVPN FDBs.
Create FDB structure and flows based on the FDB learned from the
netlink neighbor table. This is required to make the traffic unicast.
We can do the mapping between binding and remote VTEP FDB by
populating the remote FDB side table.

The flows in that table will be utilized in future commit.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>