Add initial schema and skeleton for OVN bridge controller.
This commit is the first of the series to add a new service
to OVN called ovn-br-controller (OVN bridge controller).
This commit adds
 - a schema file
 - a basic skeleton of ovn-br-controller service (which does nothing)
 - ovn-brctl utility similar to ovn-nbctl/ovn-sbctl.

Rational for adding this service:
There's a need to configure the provider bridges with specific OpenFlow
rules after packets leave the OVN pipeline and enters these provider
bridges  via the patch ports.  Since OVN has a very good abstraction
of the OpenFlow rules using OVN logical flows, we can leverage the same
here so that CMS doesn't have to deal with the specifics of OpenFlow
protocol.  Also if the CMS is written in golang or other languages,
CMS has to mostly rely on ovs-vsctl/ovs-ofctl to program the flows.
Adding this support in OVN avoids the CMS to exec these utils to
add/delete and dump the existing OpenFlow rules.

Any user can also use this new service to program and manage any OVS
bridge without using OVN and hence this service is named as
"ovn-br-controller."

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>