controller: Add support for remote VTEP learning.
In order to get the VTEP setup listeners for each special interface
that is dedicated to given VNI. Those interfaces need to be
part of the VRF, created by CMS. Once those listeners are in place
ovn-controller will receive notifications about remote VTEPs.

Create remote VTEP representation based on the data received from
netlink neighbor table. There are three key attributes that
uniquely identify each remote VTEP:
1) IP of the remote tunnel.
2) Destination port of the remote tunnel.
3) VNI.

Create a map of all remote VTEPs based on those three attributes.
The map will be used later on to crate the EVPN binding. One
thing to note is that all of this is specific to each ovn-controller.
This is one of the reasons why those data are local to each
ovn-controller, the second reason is scalability as we would have
possibly duplicate SB entries differentiated only by residing
chassis.

Note that there isn't any way to set vni for given LS that will
be part of future patch ("dynamic-routing-vni").

Acked-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>