northd, tests: Network Function insertion logical flow programming.
The objective is to insert a Network Function (NF) in the path of outbound or
inbound traffic from/to a port-group. The NF VM will be like a bump in the wire
and should not modify the packet, i.e. the IP header, the MAC addresses, VLAN
tag, sequence numbers remain unchanged.

The port-group is the point of interest here when defining the NF ports. The
“inport” is the port closer to the port-group and “outport” is the one away from it.
For from-lport ACLs, the request packets would be redirected to the NF “inport”
and for to-lport ACLs, the request packets would be redirected to NF “outport”.
Statefulness will be maintained, i.e. the response traffic will go through
the same pair of NF ports but in reverse order.

The ACL refers to a network-function-group (NFG). The NFG contains a list of
NFs. The first NF in that list is used for redirection rule.
A future patch would introduce selection of NF based on health monitoring.

If the port-group members are spread across multiple LSs, the CMS needs to
create child ports for the nf LSPs, in each LS. Northd then uses the child
ports for the redirection rules.

This patch is adding northd logic to program the logical flows in the LS.

Logical Switch ingress pipeline:
(NF port referred below can be the port configured in the NF entity or the
associated child port on the LS where the flows are being programmed.)

- in_acl_eval:
  - Existing from-lport ACL rule updated with extra action to set reg8[21]=1 to
    indicate NFG, reg8[22]=<direction> (1 - req, 0 - resp), reg5[0..7]=<NFG id>
  - Existing 65K priority rules that allow reply/related packets, updated with
    new action to copy CT label nfg bit into reg8[21] to indicate NFG case.
- in_stateful:
  - Pri 110: set NFG id in CT label if reg8[21] is set.
- in_network_function: New table
  - Pri 100: packets incoming from NF ports, simply send to next table.
  - Pri 100: multicast packets matching NF ACL, send to next table.
  - Pri 99: match NFG id (reg5) and redirect req packet to NF inport.
  - Pri 99: match NFG id (ct_label) and redirect resp packet to NF inport.
  - Pri 1: if reg8[21] set, but NF port not present on this LS, drop packet.
  - Pri 0: default forward to next table
- in_l2_lookup: Pri 100: if reg8[23] == 1 (packet has come back from egress).
                Simply forward such packets as outport is already set.

Logical Switch egress pipeline:
- out_acl_eval:
  - Existing to-lport ACL rule updated with extra action to set reg8[21]=1 to
    indicate NFG, reg8[22]=<direction> (1 - req, 0 - resp), reg5[0..7]=<NFG id>
  - Existing 65K priority rules that allow reply/related packets, updated with
    new action to copy CT label nfg bit into reg8[21] to indicate NFG case.
- out_stateful:
  - Pri 110: set NFG id in CT label if reg8[21] is set.
- out_network_function: New table.
  - Pri 100: multicast packets matching NF ACL, send to next table.
  - Priority 100: For packets going to NF ports, simply send to next table.
  - Pri 99: match NFG id (reg5) and redirect req pkt to NF outport.
  - Pri 99: match NFG id (ct_label) and redirect resp packet to NF outport.
    In both the above, redirection involves setting the packet outport, setting
    reg8[23]=1 and injecting the packet to in_l2_lookup.
  - Pri 1: if reg8[21] set, but NF port not present on this LS, drop packet.
  - Priority 0: default forward
- out_pre_acl: If packet is coming from NF inport, skip the egress pipeline upto
    the out_nf stage, as the packet has already gone through it and we don't
    want the same packet to be processed by CT again.

Two new options have been added to LSP and these are expected to be set by the
CMS (in addition to disabling port security):
lsp_learn_fdb - set to false to prevent learning MAC based on packets from NF ports.
receive_multicast - set to false to prevent forwarding of multicast traffic to NF ports.

Signed-off-by: Sragdhara Datta Chaudhuri <sragdha.chaudhu@nutanix.com>
Acked-by: Naveen Yerramneni <naveen.yerramneni@nutanix.com>
Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>