northd: Unify the the check for stale MAC Binding.
The SB MAC_Binding and Static_MAC_Binding use the same columns
as an indication of logical port and datapath. However, the check
if any of those rows is stale wasn't consistent. Unify it in a single
function to make sure that both tables have the same condition to
identify stale entries.

Fixes: 417f1415b2f5 ("northd: Clean up SB MAC bindings for deleted ports.")
Fixes: 6856adc616a7 ("ovn-northd: Clear SB records depending on stale datapaths.")
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>