ovn-sb: Adjust Datapath_Binding schema to avoid problems.
Commit 6919992d8781 changed ovn-northd to sync the NB logical
datapath UUID with the corresponding SB Datapath_Binding. They thus
would have the same UUID. This causes issues on upgrades since it
involves having to essentially recreate all southbound datapath
bindings, as well as all records that refer to those southbound datapath
bindings.

This commit changes so that instead of having the same UUID, the
southbound datapath_binding has a new nb_uuid column that contains the
UUID of the northbound logical datapath. This way, there is still a
simple way to link the Datapath_Binding to the northbound logical
datapath, but there is no need to have to update pre-existing
Datapath_Bindings to have new UUIDs.

Commit 6919992d8781 also added a new "type" field to the
Datapath_Binding table. This commit also updates the "type" field
to be optional. This is to assist with databases that are in
active-backup mode. Databases with "old" style Datapath_Binding records
would not have a "type" set. This would cause schema constraint issues
with the backup database:

  unexpected ovsdb error: constraint violation: ""
  is not one of the allowed values ([logical-router, logical-switch])

This commit changes the "type" field to be optional.

Fixes: 6919992d8781 ("Datapath_Binding: Separate type column and sync NB.UUID to SB.")
Co-authored-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>