ofctrl: Optimize and specialize the ofctrl_add_or_append_flow function.
The function has only single call site and was used for modifying
flows where the only actions in the list were conjunctions. By
specializing we can make durther assumptions that allows us to
optimize as much as possible. The main issue with the function was
very high number of allocations, 90% of them were actually to perform
a duplication check when we try to append new conjnctions into
already existing flow.

The strategy to avoid those allocations is to allocate the array
of references only once instead of going one by name effectively
turning the allocation complexity of that function fron O(N), where
N is the number of already existing conjunctions into O(1) where we
allocate constant number of times depending on the code path.

This also improves the performance of operations directly related
to conjunctions, mainly port group and address set modifications
within ACLs.

Below are numbers from a test run with 2000 LSPs, 1000 PGs, 1000 ACLs
that resulted with flows that had 1000 conjunctions as an action.

Without the change:
ovn-controller(s) completion:		910ms
node: lflow_output, handler for input port_groups took 427ms

With the change:
ovn-controller(s) completion:		407ms
node: lflow_output, handler for input port_groups took 157ms

So the speed up for the sync call is on avarage aroung 2.2x faster
and the speedup of precessing those lflows due to port group change
is on avarage 2.7x faster. This is also not limited only to port
groups, but the address set changes resulting in conjunctions will
benefit too.

Fixes: 5ad4e5395b9e ("ofctrl: Prevent conjunction duplication")
Fixes: e659bab31a91 ("Combine conjunctions with identical matches into one flow.")
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>