northd: Prevent ovn-nbctl --wait=sb from returning early.
When the engine wouldn't run for whatever reason we could end up
in a situation when SB commit would return UNCHANGED. That on its
own is completely fine, however when we are in the middle of waiting
for sync this would result in SB idl cur_cfg bump. The cur_cfg value
is used for the synchronization itself, so the ovn-nbctl call would
return early. To prevent that make sure we commit into the SB only
if engine had the chance to run. Forcing the SB Global nb_cfg update
in single transaction as the rest of the changes.

Reported-at: https://issues.redhat.com/browse/FDP-1712
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>