northd: Avoid committing DNAT traffic to SNAT zone.
Commit 40136a2f2c84 introduced the ability to directly reach networks behind
SNAT and DNAT_AND_SNAT on Distributed Routers, matching the behavior
of Gateway Routers. This was achieved by committing the incoming traffic into
the SNAT TC zone, which allowed the router to SNAT only traffic that originated
from inside the SNATed network.
An unintended consequence was that incoming traffic destined for the external
IP of a DNAT_AND_SNAT rule would also be committed to the SNAT CT zone,
resulting in an unnecessary entry, lingering in the table until it expired.

This change attempts to fix that by handling traffic matching DNAT_AND_SNAT
rule with slightly higher priority than the simple SNAT, and committing it to
DNAT CT zone instead.

Tests for the connectivity are also improved. Previously the system tests
'DNAT and SNAT on distributed router - N/S' used:
* 'ping' for ICMP traffic
* 'nc -u -z' for UDP traffic
* 'nc --send-only' for TCP traffic

It was especially the TCP traffic test that was problematic, because it would
not detect errors that occurred after the initial TCP handshake.
This patch proposes a new macro NS_CHECK_CONNECTIVITY that validates the
connectivity more thoroughly by exchanging multiple messages over the TCP
and UDP to make sure that the traffic can flow.

Fixes: 40136a2f2c84 ("northd: Fix direct access to SNAT network.")
Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>