northd: Disable resolve arp for unknown addresses in the LR.
Consider the following scenario:
(192.168.10.10/24) VM1 -> LS -> LR -> TS -> LR -> LS -> VM2 (192.168.20.20/24)

Also, LSPs from LS have the addresses configured, it's not used unknown addresses
in this case.

Ping from VM1 to VM2 works correctly, it's ok. But, if VM1 try to flood,
for example the IP 192.168.20.30 (an innexistent LSP), we'll see a high CPU load in
ovn-controller from the network destination due to the get_arp function.
For this case, ovn-controller have ovn-is-interconn = true.
To fix it, the idea could be the following logical flow in LR:
  table=22(lr_in_arp_resolve  ), priority=50   , match=(ip4.dst == 192.168.20.0/24 && reg0 != 192.168.20.1), action=(drop;)

Signed-off-by: Lucas Vargas Dias <lucas.vdias@luizalabs.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>