controller: Make sure we also match EVPN FDB on vni.
The EVPN FDB was matching only on the MAC address itself, that
would cause issues when there would be 2 EVPN fabrics but same
MAC address for a workload. Only the "last" would be learned.
Add VNI into the structure to ensure that the FDB is unique
per MAC and also per VNI.

Fixes: 3826e4071afb ("controller: Create physical flows based on the advertised EVPN FDBs.")
Reported-at: https://issues.redhat.com/browse/FDP-1730
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>