tests: Simplify "MAC binding aging - probing GW router Dynamic Neigh"
The test was checking if we are sending ARP probes when the MAC
Binding entry is close to expiration. But due to the possible race
condition between the flow and timestamp it could fail on slower
systems. Make sure we check the send part via the log message
produced by ovn-controller. That shouldn't have any race whatsoever.
At the same time slightly decrease the aging so the test doesn't take
that long to finish.

Fixes: 1e4d4409f391 ("controller: Send ARP/ND for stale mac_bindings entries.")
Reported-at: https://issues.redhat.com/browse/FDP-1732
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>