system-tests: Ensure test interfaces are always deleted.
The ADD_VETH() / ADD_VETH_BOND() macros didn't schedule deletion of the
test interfaces (by calling 'on_exit') immediately after the interfaces
were created.  That means that in the eventuality of a command failure
in between creation and the call to 'on_exit' the test interfaces
wouldn't be cleaned up.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>