controller, northd: Do not cancel recompute on SB readonly nodes.
The engine would cancel node recompute if it wasn't allowed,
typically the SB is in read-only mode. That would lead to cancel
of the node recompute and force recompute the next engine run, which
is very wasteful. Add a way to mark the engine nodes as SB write.
Only nodes that are marked as SB write will lead to cancel when
recompute is not allowed. Other nodes (SB read-only) will be able
to proceed with recompute without any issue.

To help with correct marking of the nodes there is a new IDL mechinsm
that will assert if the txn is marked as read-only but something
attempted to write into it.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>