northd: Ensure unicast ARPs are forwarded properly.
By default, northd does not install any responder flows for unicast ARP.
These are intended to be forwarded to the destination so that the
destination can respond appropriately. For VIF ports, this tends to work
as expected in most scenarios.

When proxy ARP is configured on a logical switch port conntected to a
logical router, then we install low priority flows to ensure that
ARPs for the configured proxy addresses is responded to by the logical
switch. These proxy ARP flows are hit when unicast ARP requests are sent
for the VIF ports on the logical switch. We therefore end up responding
incorrectly to unicast ARP requests with the proxy ARP MAC instead of
forwarding the ARP request to the proper VIF port.

This commit fixes the issue by installing explicit "next;" actions for
unicast ARP requests directed towards VIFs so that the proxy ARP flows
will not be hit. These flows are only installed if proxy ARP is
configured, since they are unnecessary otherwise.

Reported-at: https://issues.redhat.com/browse/FDP-1646
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>