ci: ovn-kubernetes: Enable "External Gateway" ovn-kubernetes CI lanes.
Since [0] some feature tests have been moved out of the default
"control-plane" CI lane in ovn-kubernetes.  In order to run those, they
have to be explicitly triggered.  When that change happened the OVN CI
lanes were not updated so we lost test coverage for those features.

This lead to regressions like [1] creeping in.

This commit re-enables the tests.

[0] https://github.com/ovn-kubernetes/ovn-kubernetes/commit/cf116ea
[1] https://mail.openvswitch.org/pipermail/ovs-dev/2025-September/426000.html

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>