tests: Skip "Load Balancer LS hairpin IPv6 UDP - larger than MTU"
This test currently fails with certain kernel versions. It has been seen
to fail with the following:

Fedora 41 running 6.14.9-200.fc41.x86_64
Ubuntu 24.04 running 6.14.0-1012-azure

The working theory is that these kernels have an improperly backported
change that results in all IPv6 UDP communications disabling
fragmentation. For our test that specifically sends IPv6 UDP packets
larger than the MTU, this causes the test to fail unexpectedly.

This commit is intended to allow the test to be skipped temporarily in
Github CI runs until the runner image is updated to have the proper
backport.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>