northd: Add disable arp and nd resolve option for logical_router_port.
Consider the following scenario:
(192.168.10.10/24) VM1 -> LS -> LR -> TS -> LR -> LS -> VM2 (192.168.20.20/24)

Also, LSPs from LS have the addresses configured, it's not used unknown addresses
in this case.

Ping from VM1 to VM2 works correctly, it's ok. But, if VM1 try to flood,
for example the IP 192.168.20.30 (an innexistent LSP), we'll see a high CPU load in
ovn-controller from the network destination due to the get_arp function.
For this case, ovn-controller have ovn-is-interconn = true.

In the connection LR -> TS -> LR all addresses are known, so arp resolve
could be disable. For this case, add the option disable_arp_resolve=true
to logical_router_port drop the packets for unknown addresses. If the
communication is IPv6, configure the option disable_nd_resolve=true.

Co-authored-by: Tiago Matos Carvalho Reis <tiago.reis@luizalabs.com>
Signed-off-by: Tiago Matos Carvalho Reis <tiago.reis@luizalabs.com>
Signed-off-by: Lucas Vargas Dias <lucas.vdias@luizalabs.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>