expr: Use prefixes instead of single bit masks for inequality.
While converting inequality matches into sets of equality matches
we use the most simple and efficient method - create single bit masks
matching on the opposite value of each bit.  For example, 'a != 0b0101'
where 'a' is a 4-bit field turns into (note, a[0] is the least
significant bit, a[3] is the most significant):

  a[0] != 1 || a[1] != 0 || a[2] != 1 || a[3] != 0
or
  a[0] == 0 || a[1] == 1 || a[2] == 0 || a[3] == 1

This corresponds to the following matches:

           LSB    MSB
              0123
  original :  1010
  a[0] == 0:  0...
  a[1] == 1:  .1..
  a[2] == 0:  ..0.
  a[3] == 1:  ...1

We can technically preserve part of the mask that is already covered
by previous matches like this:

  a[0..3] == 0b0100 || a[1..3] == 0b011 || a[2..3] == 0b00 || a[3] == 1

Which just corresponds to filling the upper half of the matrix with
the original values:

           LSB    MSB                   LSB    MSB
              0123                         0123
  original :  1010                         1010
  a[0] == 0:  0...     a[0..3] == 0b0100:  0010
  a[1] == 1:  .1..     a[1..3] == 0b011 :  .110
  a[2] == 0:  ..0.     a[2..3] == 0b00  :  ..00
  a[3] == 1:  ...1     a[3]    == 0b1   :  ...1

The diagonal values are the same in both sets of matches.

Both sets of matches are logically equivalent and fully represent the
original inequality match.

This is true, because if the value 'a' equals 0b0101, then it will not
match any of the rules in the set, as each of them has at least one
bit different from 0b0101 (diagonal ones).  And if the value 'a' is
different from 0b0101, then we can find the first bit that is different
starting from the most significant, all the higher ones will be the
same as in 0b0101.  And we have a rule that matches on exactly that by
design.

Now, since every match in the second set has a contiguous mask ending
at the most significant bit, we can re-write it using a CIDR-like
notation:

  a == 0b0100/4
  a == 0b0110/3
  a == 0b0000/2
  a == 0b1000/1

Or: a == 0b0100/4 || a == 0b0110/3 || a == 0b0000/2 || a == 0b1000/1

Why does it matter?

Having larger masks allows much easier elimination of incompatible
sub-expressions during the normalization process.

Let's say we have the following expression:

 a != 0b1000/2 && a != 0b0101/4

Let's normalize it by swapping the inequality with equality sets using
the single-bit method:

 (a[2] == 1 || a[3] == 0) &&
 (a[0] == 0 || a[1] == 1 || a[2] == 0 || a[3] == 1)

Opening the parenthesis:

 a[2] == 1 && a[0] == 0 ||
 a[2] == 1 && a[1] == 1 ||
 a[2] == 1 && a[2] == 0 ||   < Inconsistent
 a[2] == 1 && a[3] == 1 ||
 a[3] == 0 && a[0] == 0 ||
 a[3] == 0 && a[1] == 1 ||
 a[3] == 0 && a[2] == 0 ||
 a[3] == 0 && a[3] == 1      < Inconsistent

We can see that two sub-expressions can be eliminated as they are
always false.  Only the sub-expressions that match on the same bit
can be eliminated at this stage, the rest must be processed further.

With the prefix method we get:

 (a[2..3] == 0b11 || a[3] == 0) &&
 (a[0..3] == 0b0100 || a[1..3] == 0b011 || a[2..3] == 0b00 || a[3] == 1)

In a bit easier to read form:

 (a == ..11 || a == ...0) &&
 (a == 0010 || a == .110 || a == ..00 || a == ...1)

Opening the parenthesis:

   LSB    MSB     LSB    MSB
      0123           0123
 a == ..11  &&  a == 0010 ||  < Inconsistent
 a == ..11  &&  a == .110 ||  < Inconsistent
 a == ..11  &&  a == ..00 ||  < Inconsistent
 a == ..11  &&  a == ...1 ||
 a == ...0  &&  a == 0010 ||
 a == ...0  &&  a == .110 ||
 a == ...0  &&  a == ..00 ||
 a == ...0  &&  a == ...1     < Inconsistent

Now four sub-expressions are always false instead of just two.  And we
can see that if we had more bits matched in the second inequality
match instead of 4, then all the rest of 'a == ..11 && ...'
sub-expressions would also be always false.  By not dealing with all
possible bit masks and reducing the problem to a much smaller space
of prefix matches, we're significantly increasing chances of resulted
sub-expressions to end up internally inconsistent and discarded.

This greatly reduces the amount of sub-expressions that we need to
process on the next steps.  Removal of duplicates and crushing of
supersets will get rid of most of the unnecessary matches in the case
of single-bit method, but not all of them, since we do not have a lot
of context to track dependencies between different sub-expressions.
And it's also much more expensive to do so, since superset crushing
has quadratic complexity in the worst case, which grows pretty fast
with the amount of input sub-expressions.  More we can eliminate
before this stage, the better.

In practice, using the prefix method provides a huge performance
boost while processing inequality matches.  For example, normalizing
the inequality match on the following set of 15 random IP addresses
takes about 28 minutes with a single bit approach, generating 60519
OpenFlow rules:

  ip4.dst != {
    179.141.79.238/32,  23.87.193.64/32,   240.238.112.253/32,
    42.140.78.76/32,    243.50.68.23/32,   229.151.116.21/32,
    201.156.140.194/32, 94.49.119.92/32,   133.167.173.49/32,
    51.199.153.252/32,  214.80.176.168/32, 134.195.2.51/32,
    68.18.121.185/32,   118.53.215.77/32,  203.42.132.181/32,
  }

With the prefixes, the same task takes just about 4 milliseconds,
generating only 412 OpenFlow rules.  Generated rules match the full
negation of the set performed with python netaddr library:

  from netaddr import IPSet
  list((IPSet(['0.0.0.0/0']) - IPSet(exclusions)).iter_cidrs())

Additional bonus is that prefix matches are much friendlier to the
OpenFlow classifier in OVS that can now utilize the prefix match
optimizations for some of these rules, potentially reducing the amount
of datapath flows.

The superset crushing logic is still needed to get rid of the extra
rules when processing sets with particularly diverse masks and
custom bit matches, so keeping it as it is.  New tests added to
substitute the reduced test coverage for this functionality.

The prefixes approach is still not a silver bullet and it will take
noticeable amount of time on very large sets of random IPs with
inequality matches, but we're moving the scalability pole way further
with a relatively minor change.  For example, it will take 2.7 seconds
to normalize inequality match on a set with 200 random /32 addresses
and 38 seconds on a set with 500 random /32 addresses.  In real world
applications the speed should be much higher as IPs are normally not
that random.  With the single-bit method, normalization of such sets
is practically impossible.

If we'll need further improvements for handling very large sets, we'll
likely need to special-case the inequality checks for constant sets in
curly braces that have CIDR-like masks.  This way we could negate the
set as a whole, which is much simpler computationally, but a fair bit
heavier code-wise.  This approach would not handle expressions without
curly-braced sets or with multiples of them and will not handle
different types of masks, while generic conversion of each negation
into a set of prefix matches doesn't require any specific format of
the original expression and works for any masks, even if they are not
contiguous.

Reported-at: https://issues.redhat.com/browse/FDP-1952
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>