tests: Handle change of behavior in iproute2.
The iproute2 tool recently changed behavior [0] which has effect
on representation of values currently used for assertions in the
"egress qos" system test.

The purpose of the OVN test case is to ensure the system is
updated on the back of changes to the database and not to
implementation details of iproute2.

Relax the assertions made on the tc representation of htb cburst
values.

0: https://git.kernel.org/pub/scm/network/iproute2/iproute2.git/commit/?id=d947f365602b30657d1b797e7464000d0ab88d5a
Reported-at: https://bugs.launchpad.net/ubuntu/+source/iproute2/+bug/2129005
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>