ovn: Fix SNI support in a few utilities.
Commit af7c9802cb (ovn: Add SNI support to ovn-nbctl and ovn-sbctl.)
intended to add SNI support for ovn-nbctl and ovn-sbctl explicity.
There was implicit support to some utilties using
STREAM_SSL_OPTION_HANDLERS (ovn-controller-vtep, ovn-ic-nbctl,
ovn-ic-sbctl, ovn-trace).

But it missed adding explict support to utilities ovn-ic, ovn-northd
and ovn-controller causing those utilities
to crash if --ssl-server-name was passed as an argument.

This commit fixes it and also adds a NEWS item for the feature.

CC: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Gurucharan Shetty <guru@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>