controller: Prevent transaction error for BGP routes.
When ovn-controller tried to learn two routes with the same
destination and next-hop, but different metric, it would end up
in a transaction error and loop. Make sure we are also checking
routes that are learned in the current I-P run to prevent that
issue.

Fixes: 866a5014ae45 ("controller: Support learning routes.")
Reported-at: https://issues.redhat.com/browse/FDP-2081
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>