northd: Fix the lflow_handle_ls_stateful_changes function.
There were two issues with the handler:
1) The handler was handling deleted ls_stateful, but in a wrong way.
   It was trying to find ovn_datapath struct which was already
   removed by northd.

2) The handler tried to sync the flows one by one which was causing
   performance issues as the dp groups were created over and over
   again.

None of the above was an issue because we would recompute the whole
lflow nodes in the both cases above. This was noticed with the
attempt to introduce incremental processing for LS addition.

To fix the issues above make sure we don't check the ovn_datapath for
deleted ls_stateful struct. Just resync the logical flows. Also make
a separate loop that will go over the ls_stateful updated structs
once again just to sync after all datapath groups are actually up to
date.

Finally make sure we propagate the updated correctly for newly
created ls_stateful structs.

Fixes: 00cda5a47bd4 ("northd: Add ls_stateful handler for lflow engine node.")
Fixes: 48e1736f89c6 ("northd: I-P for logical switch creation/deletion in en_ls_stateful engine node.")
Co-authored-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>