Revert "northd: Add and delete logical routers in en-lflow engine node.".
Do a partial revert of the I-P handling for LR addition and deletion.
This causes issues with DP groups processing, when there new flows
added northd is still holding multiple references to the existing DP
group. In that case we cannot resize it, because we don't know if the
DP group will split. This might lead to DP groups being deleted and
recreated as we slowly remove the references and at the end all flows
migrated to the new DP group leaving the old one empty. This is
particularly bad for SB performance as we are hammering it with DP
group creation/deletion + updates of DP group references in all
affected flows.

The lflows_ref is intentionally left NULL in the ovn_datapath struct
so we don't have to revert all those changes to pass it along, and
so we don't waste any space to maintain them as they are not used
by anything right now.

This reverts commit 9ec96d0d85b66a6d2c9e66da694c9cfe710bdf46.
This reverts commit cf92fb76ed487862054b8d10c89a46f5fb796072.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>