datapaths: Handle simultaneous "new" and "deleted" datapaths.
If a logical switch or logical router is added and removed very quickly,
then it is possible for the IDL to place it in both the "new" and
"deleted" tracked maps.

This was not being accounted for in the logical switch and logical
router handlers.

For logical routers, this was not a huge issue. We would detect the
logical router as deleted, then not be able to find the in-memory
unsynced datapath. This would result in a recompute. Adding the check
for this specific situation allows us to avoid an unnecessary recompute.

For logical switches, on the other hand, this was a much worse
situation. The code first checks if the logical switch is new. In this
case, the added and deleted logical switch matched this check, so we
treated it as a new logical switch. This "new" logical switch would then
get propagated to all other engine nodes. If this logical switch
happened to reference other database rows that also had been deleted,
this could cause chaos. One place in particular where this was observed
was in the en_lb_data node. If the logical switch referenced a load
balancer group that was deleted, then we would try to look up the load
balancer group in our in-memory map of load balancer groups. However,
since the actual database row had been deleted, it no longer existed in
our in-memory map. This resulted in an assertion being triggered and
crashing ovn-northd.

Added and deleted logical switches could also trigger a crash in the
datapath syncing code as well. Since the logical switch was deleted, its
memory could be freed by the IDL. This could cause a crash at any point
in its processing.

The fix here is to detect the situation where the IDL reports the
logical router or logical switch as both new and deleted. In this case,
we treat it as a no-op and continue. This avoids both the recompute for
logical routers, and the crashes with logical switches.

Reported-at: https://issues.redhat.com/browse/FDP-2780
Reported-at: https://issues.redhat.com/browse/FDP-2805
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>