build: Require `which` to build.
`find_command` test helper assumes `which` is present on the system. If
it is not, the helper silently fails to detect an installed command
presence, skipping tests that a caller could expect to run (since the
auxiliary commands *are* installed).

Other build scripts and test cases also use the command, and we install
it in the fedora container too.

This patch enforces `which` presence via autoconf.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>