multinode.at: Test OVN EVPN integration for IPv6 overlays.
This extends the existing multinode EVPN tests to make sure the feature
also works fine for IPv6 logical networks.

NOTE: ideally we should also test with IPv6 underlays (IPv6 VXLAN tunnel
endpoints) but that's not possible yet because we rely on FRR for the
BGP control plane and FRR doesn't support IPv6 VTEPs in any of the
available releases.  In the future that should be possible though,
once [0] is included in an official FRR release.

[0] https://github.com/FRRouting/frr/pull/19498

Reported-at: https://issues.redhat.com/browse/FDP-2771
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>