datapaths: Create enum alias for minimum datapath type.
This is a cosmetic change that establishes a clearer way to iterate over
all known OVN datapath types. Some for loops started iterating with "0",
and others started with "DP_SWITCH" since it was the first value in the
enum. Neither of these is great. 0 is not an enum value in
ovn_datpaath_type. DP_SWITCH is not a clear way of indicating the first
value in the enum, and it's vulnerable to enum changes that might not
keep DP_SWITCH as the first value.

This change introduces "DP_MIN" as the first value in the enum, and
aliases DP_SWITCH to be the same as DP_MIN. This way, for loops can
start with DP_MIN instead of 0 or DP_SWITCH.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>