northd: Pass ovn_synced_datapath to lflow addition macros.
The previous patches in this series have made it so that ovn_datapath is
no longer required for managing lflow tables or syncing lflow tables to
the southbound database. The only remnant is the helper macros for
adding lflows to the lflow table.

This patch converts the helper macros to take an ovn_synced_datapath
instead of an ovn_datapath. This also converts all uses of these macros
in northd.c

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>