lib/actions: Check for tunnel_ofport fn ptr null before calling.
This check is added to address NULL deference issue reported by
coverity.  br-controller also uses the lib/actions and it sets
the 'tunnel_ofport' in the 'struct ovnact_encode_params' to NULL.
br-controller can crash if a user uses the OVN logical action -
fwd_group() even though it is not supported.

Ideally br-controller should define its own supported OVN actions
as a proper fix.  This patch is just a temporary fix until we add
this support.

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>